logging: reorganize files, define protobuf configuration/control
