+++ datatypes/Logging.proto
@@ -1,0 +1,31 @@
+syntax = "proto3";
+
+package MRA.Datatypes;
+
+enum LogLevel // needs to be consistent with levels.hpp
+{
+    CRITICAL = 0;
+    ERROR = 1;
+    WARNING = 2;
+    INFO = 3;
+    DEBUG = 4;
+    TRACE = 5;
+}
+
+message LogSpec
+{
+    string component = 1; // name of the component, for example "robotsports/proof_is_alive"
+    LogLevel level = 2; // level to use, default INFO
+    bool toFile = 3; // write data to file, default true
+    bool dumpTicks = 4; // write binary protobuf data to a file per tick, default false - can hurt performance but is valuable for debugging and unittesting
+    int32 maxLineSize = 5; // number of characters to cut off lines, default something like 1000 - if this is too little, then consider using tickdump
+    float maxFileSizeMB = 6; // a guard on file size growth, default something like: stop after file exceeds 10.0 MB
+}
+
+message LogControl
+{
+    string folder = 1; // folder to write to, default auto-create in /tmp, something like /tmp/mralog_20230826_090811
+    LogSpec general = 2; // settings to use if no component overrule is specified
+    repeated LogSpec overrules = 3; // component overrules, when not specified, use general settings
+}
+
