+++ libraries/logging/backend.cpp
@@ -1,4 +1,4 @@
-#include "MRA_logger.hpp"
+#include "backend.hpp"
 #include <memory>
 #include "spdlog/spdlog.h"  // spdlog API: https://github.com/gabime/spdlog
 #include "spdlog/async.h"
