+++ libraries/logging/levels.hpp
@@ -1,0 +1,15 @@
+#ifndef _MRA_LIBRARIES_LOGGING_LEVELS_HPP
+#define _MRA_LIBRARIES_LOGGING_LEVELS_HPP
+
+enum LogLevel
+{
+    CRITICAL,
+    ERROR,
+    WARNING,
+    INFO,
+    DEBUG,
+    TRACE
+};
+
+#endif // #ifndef _MRA_LIBRARIES_LOGGING_LEVELS_HPP
+
