--- libraries/logging/logging.cpp
@@ -1,4 +1,0 @@
-#include "logging.hpp"
-
-
-
